Create Anadrome
An anagram is a word or phrase formed by rearranging the letters of a different word or phrase. For example, "Sool" is an anagram for "Solo". A palindrome is a word or phrase which reads the same backward as forward, such as "madam". An anadrome is a word or phrase if any of its anagrams form a palindrome. 
For example: 
Input: "SoloSolo" 
Output: yes ("SoollooS" is an anagram of "SoloSolo" which also is a palindrome). 
Input: "3haha" 
Output: yes ("ha3ah" is an anagram of "3haha" which also is a palindrome). 
Input: "Solo" 
Output: no 
Write a program to check if the user input is an anadrome or not.